Revert "Don't remove chat counter on game end"
This reverts commit ad8621411b58403201dba007d28f38af7e9fe0f0.
This is for coherence between lichess games. If we get back to a
finished game that was played on lichess.org there won't be any storage
id associated to the game (contrary to a game played on mobile).

Thus, the behavior is the same for a game played on lichess.org or on
mobile app.