Revert "Don't remove chat counter on game end"
This reverts commit 53c0b3bca43edf7d1ae4fe1410395e7da677f206.
This is for coherence between lichess games. If we get back to a
finished game that was played on lichess.org there won't be any storage
id associated to the game (contrary to a game played on mobile).

Thus, the behavior is the same for a game played on lichess.org or on
mobile app.